role and permission

+++ app/Http/Controllers/Backend/CategoryController.php
@@ -6,6 +6,8 @@ use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
 
 use App\Models\Category;
+use Illuminate\Support\Facades\Auth;
+use Symfony\Component\HttpKernel\Exception\HttpException;
 
 class CategoryController extends Controller
 {
@@ -24,6 +26,10 @@ class CategoryController extends Controller
      */
     public function create()
     {
+        // Check if the user is an admin
+        if (!Auth::check() || !Auth::user()->hasRole('admin')) {
+            throw new HttpException(403, 'You do not have permission to access this resource.');
+        }
         return view('backend.category.create');
     }
 
@@ -55,6 +61,10 @@ class CategoryController extends Controller
      */
     public function edit(string $id)
     {
+        // Check if the user is an admin
+        if (!Auth::check() || !Auth::user()->hasRole('admin')) {
+            throw new HttpException(403, 'You do not have permission to access this resource.');
+        }
         $category = Category::where('id', $id)->first();
         return view('backend.category.edit', compact('category'));
     }
@@ -64,6 +74,11 @@ class CategoryController extends Controller
      */
     public function update(Request $request, string $id)
     {
+        // Check if the user is an admin
+        if (!Auth::check() || !Auth::user()->hasRole('admin')) {
+            throw new HttpException(403, 'You do not have permission to access this resource.');
+        }
+
         $category = Category::find($id);
 
         $category->name = $request->input('name');
@@ -80,6 +95,11 @@ class CategoryController extends Controller
      */
     public function destroy(string $id)
     {
+        // Check if the user is an admin
+        if (!Auth::check() || !Auth::user()->hasRole('admin')) {
+            throw new HttpException(403, 'You do not have permission to access this resource.');
+        }
+
         $category = Category::find($id);
         $category->delete();
